backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
asyncpg==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0
transformers==4.41.0
torch==2.3.0
sentence-transformers==3.0.0
detoxify==0.5.2
openai-whisper==20231117
librosa==0.10.2
soundfile==0.12.1
pytesseract==0.3.10
Pillow==10.3.0
nudenet==3.4.2
reportlab==4.2.0
python-multipart==0.0.9
numpy==1.26.4
scikit-learn==1.5.0
datasets==2.19.1
evaluate==0.4.2
accelerate==0.30.0
networkx==3.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.3
qrcode==7.4.2
emoji==2.12.1
langdetect==1.0.9
sentencepiece==0.2.0
sacremoses==0.1.1
# IndicTrans2 — install separately:
# pip install git+https://github.com/AI4Bharat/IndicTrans2
# ai4bharat-transliteration — install separately:
# pip install ai4bharat-transliteration
